documents/workers/jobs/scripts/services/demo/requirements.txt
Django==2.1
django-cors-headers==3.4.0
django-rest-multiple-models==2.1.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
mysqlclient==1.4.6
PyJWT==1.7.1
pytz==2021.1
